+++ src/fides/api/models/privacy_request/__init__.py
@@ -10,13 +10,12 @@ from .privacy_request import (
     PrivacyRequest,
     PrivacyRequestError,
     PrivacyRequestNotifications,
-    RequestTask,
-    TraversalDetails,
     generate_request_callback_pre_approval_jwe,
     generate_request_callback_resume_jwe,
     generate_request_task_callback_jwe,
 )
 from .provided_identity import ProvidedIdentity, ProvidedIdentityType
+from .request_task import RequestTask, TraversalDetails
 
 __all__ = [
     "Consent",
